pass parameter file into slick_run
