+++ slick_frontend.sh
@@ -4,7 +4,7 @@ parent_path=$( cd "$(dirname "${BASH_SOURCE[0]}")" ; pwd -P )
 python "$parent_path/slick_init.py" $1
 if [ $? = 0 ]
 then
-  srun bash slick_run_jobscript.sh
+  srun bash slick_run_jobscript.sh $1
   echo "Slick has finished"
   exit 0
 else
